Changes to make compatible with latest Elemental
These are changes to make compatible with the latest Elemental trunk.
This required changing the include paths to the new elemental layout and
renaming ResizeTo to Resize.  I ended up having some of Clique's
ResizeTo's changed to Resize's, but I do not think I caught them all.

Also the examples did not have NestedDissection ifdef'd for parmetis.
Since we are interested in building without parmetis, the ifdef's were
added.